TestCaseEvalulation/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
crewai>=0.28.0
openai>=1.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openpyxl>=3.1.2
pandas>=2.0.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
python-multipart>=0.0.9
